refactor(landing): hoist shared preview bits, align redirect helper
- Move PreviewMetric/CURRENCY_FORMATTER out of MortgagePreview.tsx into
  their own module — PurchaseCostPreview.tsx was reaching sideways into
  a sibling tab's file to use them
- EvaluationSection's sign-up/log-in links now use the same dynamic
  getRedirectPath() pattern already used in HiddenCostsCalculator,
  instead of a hardcoded path
